requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asteroid-filterbanks==0.4.0
attrs==25.4.0
av==15.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
ctranslate2==4.6.2
cycler==0.12.1
deepgram-sdk==5.3.0
dnspython==2.8.0
docopt==0.6.2
dotenv==0.9.9
ecdsa==0.19.1
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.125.0
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.11
intel-openmp==2021.4.0
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.9
lightning==2.6.0
lightning-utilities==0.15.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
nltk==3.9.2
omegaconf==2.3.0
openpyxl==3.1.5
optuna==4.6.0
packaging==25.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pillow==12.0.0
primePy==1.3
propcache==0.4.1
protobuf==6.33.2
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semver==3.0.4
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.45
starlette==0.50.0
tabulate==0.9.0
tbb==2021.13.1
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.3
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
websockets==15.0.1
whisperx @ git+https://github.com/m-bain/whisperx.git@d32ec3e3012ec4c0934f4088424c32f3f038b249
yarl==1.22.0
numpy
onnxruntime
pandas
psycopg2-binary